requirements.txt
streamlit>=1.25.0
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.2.0
plotly>=5.10.0
pydeck>=0.8.0
xlsxwriter>=3.0.0
Pillow>=9.0.0
requests>=2.28.0